requirements.txt
alembic==1.4.0
amqp==2.5.2
attrs==19.3.0
backcall==0.1.0
billiard==3.6.3.0
bleach==3.1.0
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.1
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kombu==4.6.8
Mako==1.1.1
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.0.2
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
PyMySQL==0.9.3
pyrsistent==0.15.6
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
scikit-learn==0.22
scipy==1.4.1
Send2Trash==1.5.0
six==1.13.0
SQLAlchemy==1.3.13
stomp.py==4.1.22
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
zipp==0.6.0
